[accel-wave] refactor directories
